Refactor code in smaller files and remove most IDE warnings
